+++ Frends.MongoDB.Delete/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [2.0.0] - 2024-12-05
+### Updated
+- Updated MongoDB.Driver to version 3.1.0
+
 ## [1.0.1] - 2023-11-22
 ### Fixed
 - Fixed dll error when importing the Task to Frends by adding local dll reference to the project file.