+++ Frends.MongoDB.Query/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [2.0.0] - 2024-12-05
+### Updated
+- Updated MongoDB.Driver to version 3.1.0
+
+### Added
+- Option to query single or multiple documents
+
 ## [1.0.1] - 2023-11-23
 ### Fixed
 - Fixed dll error when importing the Task to Frends by adding local dll reference to the project file.