+++ Frends.MongoDB.Update/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [2.0.0] - 2024-12-05
+### Added
+- Option to use upsert when updating documents
+
+### Updated
+- Updated MongoDB.Driver to version 3.1.0
+
+
 ## [1.0.1] - 2023-11-23
 ### Fixed
 - Fixed dll error when importing the Task to Frends by adding local dll reference to the project file.